requirements.txt
# =====================================================
# REQUIREMENTS.TXT CHO PROJECT BACKEND
# =====================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Image Processing
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging
structlog==23.2.0

# CORS
fastapi-cors==0.0.6

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Upload
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0
